dont autocast rope embeddings (#1304)
The TE rope kernels can accept an fp32 rope embedding input, so we don't
have to cast these to the hidden state's input dtype

+++ bionemo-recipes/models/llama3/modeling_llama_te.py
@@ -269,4 +269,4 @@ class NVLlamaRotaryEmbedding(LlamaRotaryEmbedding):
             freqs = (inv_freq_expanded.float() @ position_ids_expanded.float()).transpose(1, 2)
             emb = torch.cat((freqs, freqs), dim=-1)
 
-        return emb.to(dtype=x.dtype).transpose(0, 1).unsqueeze(1)
+        return emb.transpose(0, 1).unsqueeze(1)
